docs(README): update demo to Stackblitz example
The current Plunkr demo is broken. This is a known problem due to a bad initial configuration of the Angular template on Plunker which did not "freeze" the Angular version that's being fetched from unpkg.com. As a result, changes in the package structure broke a lot of demos.

I created a new demo and ported it to Stackblitz. It's much easier to update the dependencies there (just click the refresh button on the deps panel). And people even have the possibility to export it to a working Angular CLI setup.

You may want to fork my stackblitz and adjust the URL to be able to update it in the future.